Move flow declarations and test into recompose
