+++ .flowconfig
@@ -1,9 +1,9 @@
 [ignore]
+<PROJECT_ROOT>/types/.*
 
 [include]
 
 [libs]
-../flow-typed/recompose_v0.24.x/flow_v0.55.x-/recompose_v0.24.x.js
 
 [options]
 suppress_comment=\\(.\\|\n\\)*\\$ExpectError
