+++ src/packages/recompose/__tests__/types/test_mapProps.js
@@ -2,9 +2,9 @@
 /* eslint-disable no-unused-vars, no-unused-expressions, arrow-body-style */
 /* @flow */
 import React from 'react'
-import { compose, mapProps, withProps } from 'recompose'
+import { compose, mapProps, withProps } from '../..'
 
-import type { HOC } from 'recompose'
+import type { HOC } from '../..'
 
 type EnhancedCompProps = { eA: 1 }
