+++ src/packages/recompose/__tests__/types/test_mapPropsStream.js
@@ -3,9 +3,9 @@
 /* @flow */
 import React from 'react'
 // import { Observable } from 'rxjs'
-import { compose, mapProps, withProps, mapPropsStream } from 'recompose'
+import { compose, mapProps, withProps, mapPropsStream } from '../..'
 
-import type { HOC } from 'recompose'
+import type { HOC } from '../..'
 
 type EnhancedCompProps = { eA: 1 }
