+++ src/packages/recompose/__tests__/types/test_onlyUpdateForKeys.js
@@ -1,9 +1,9 @@
 /* eslint-disable no-unused-vars, no-unused-expressions */
 /* @flow */
 import React from 'react'
-import { compose, withProps, toClass } from 'recompose'
+import { compose, withProps, onlyUpdateForKeys } from '../..'
 
-import type { HOC } from 'recompose'
+import type { HOC } from '../..'
 
 type EnhancedCompProps = { eA: 1 }
 
@@ -17,7 +17,7 @@ const Comp = ({ eA }) =>
   </div>
 
 const enhacer: HOC<*, EnhancedCompProps> = compose(
-  toClass,
+  onlyUpdateForKeys(['eA']),
   withProps(props => ({
     eA: (props.eA: number),
     // $ExpectError eA nor any nor string
@@ -27,6 +27,11 @@ const enhacer: HOC<*, EnhancedCompProps> = compose(
     // $ExpectError property not found
     err: props.iMNotExists,
   }))
+)
+
+const enhacerErr: HOC<*, EnhancedCompProps> = compose(
+  // $ExpectError property not found
+  onlyUpdateForKeys(['eB'])
 )
 
 const EnhancedComponent = enhacer(Comp)
