+++ src/packages/recompose/__tests__/types/test_shouldUpdate.js
@@ -1,19 +1,9 @@
 /* eslint-disable no-unused-vars, no-unused-expressions, arrow-body-style */
 /* @flow */
 import React from 'react'
-import {
-  compose,
-  withProps,
-  setStatic,
-  setPropTypes,
-  setDisplayName,
-} from 'recompose'
-// import PropTypes from 'prop-types'
-import type { HOC } from 'recompose'
+import { compose, withProps, shouldUpdate } from '../..'
 
-const PropTypes = {
-  string: () => {},
-}
+import type { HOC } from '../..'
 
 type EnhancedCompProps = { eA: 1 }
 
@@ -27,11 +17,14 @@ const Comp = ({ eA }) =>
   </div>
 
 const enhacer: HOC<*, EnhancedCompProps> = compose(
-  setStatic('hello', 'world'),
-  setPropTypes({
-    a: PropTypes.string,
+  shouldUpdate((props, nextProps) => {
+    // $ExpectError eA nor any nor string
+    ;(props.eA: string)
+    // $ExpectError eA nor any nor string
+    ;(nextProps.eA: string)
+
+    return props.eA === nextProps.eA
   }),
-  setDisplayName('hello'),
   withProps(props => ({
     eA: (props.eA: number),
     // $ExpectError eA nor any nor string
@@ -43,13 +36,11 @@ const enhacer: HOC<*, EnhancedCompProps> = compose(
   }))
 )
 
-// $ExpectError name is string
-setDisplayName(1)
-
-// $ExpectError propTypes is object
-setPropTypes(1)
-
-// $ExpectError name is string
-setStatic(1, 'world')
+const enhacerErr: HOC<*, EnhancedCompProps> = compose(
+  shouldUpdate(() => {
+    // $ExpectError must be boolean
+    return 1
+  })
+)
 
 const EnhancedComponent = enhacer(Comp)
