+++ src/packages/recompose/__tests__/types/test_utils.js
@@ -2,9 +2,9 @@
 /* @flow */
 
 import React from 'react'
-import { compose, withProps, hoistStatics } from 'recompose'
+import { compose, withProps, hoistStatics } from '../..'
 
-import type { HOC } from 'recompose'
+import type { HOC } from '../..'
 
 type EnhancedCompProps = { a: number }
