+++ src/packages/recompose/__tests__/types/test_withContext.js
@@ -1,14 +1,9 @@
 /* eslint-disable no-unused-vars, no-unused-expressions, arrow-body-style */
 /* @flow */
 import React from 'react'
-import { compose, withProps, getContext } from 'recompose'
-// import PropTypes from 'prop-types'
-import type { HOC } from 'recompose'
+import { compose, withProps, withContext } from '../..'
 
-const PropTypes = {
-  number: () => {},
-  string: () => {},
-}
+import type { HOC } from '../..'
 
 type EnhancedCompProps = { eA: 1 }
 
@@ -22,19 +17,15 @@ const Comp = ({ eA }) =>
   </div>
 
 const enhacer: HOC<*, EnhancedCompProps> = compose(
-  getContext({
-    // as an idea is to use a hack like this
-    // so we can test all such types
-    color: ((PropTypes.string: any): string),
-    num: ((PropTypes.number: any): number),
+  withContext({}, props => {
+    // $ExpectError eA nor any nor string
+    ;(props.eA: string)
+    return {}
   }),
   withProps(props => ({
     eA: (props.eA: number),
-    color: (props.color: string),
     // $ExpectError eA nor any nor string
     eAErr: (props.eA: string),
-    // $ExpectError color nor any nor number
-    colorErr: (props.color: number),
   })),
   withProps(props => ({
     // $ExpectError property not found
