+++ src/packages/recompose/__tests__/types/test_withPropsOnChange.js
@@ -1,9 +1,9 @@
 /* eslint-disable no-unused-vars, no-unused-expressions, arrow-body-style */
 /* @flow */
 import React from 'react'
-import { compose, withProps, withPropsOnChange } from 'recompose'
+import { compose, withProps, withPropsOnChange } from '../..'
 
-import type { HOC } from 'recompose'
+import type { HOC } from '../..'
 
 type EnhancedCompProps = { eA: 1 }
